feature/3: add delete button and erase user from db
