+++ package.json
@@ -12,6 +12,7 @@
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.3.0",
     "react-scripts": "5.0.1",
+    "sweetalert2": "^11.4.19",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
