+++ src/pages/Home.css
@@ -10,6 +10,7 @@
     align-items: center;
     justify-content: center;
     margin-top: 30px;
+    margin-bottom: 30px;
 }
 
 .form input {
