+++ src/pages/Home.js
@@ -1,28 +1,23 @@
 import React, { useState, useEffect } from 'react'
 import Axios from 'axios'
-import { Form } from '../components/Form'
-import { Table, TableComponent } from '../components/Table'
+import { UsersTable } from '../components/UsersTable'
 import './Home.css'
+import { UserForm } from '../components/UserForm'
 
 export const Home = () => {
   const [usersList, setUsersList] = useState([]);
 
   useEffect(() => {
-  
   Axios.get('http://localhost:3001/readUsers').then((response) => {
       setUsersList(response.data)
     })
-  }, [])
+  }, [usersList])
   
-    
-  
-
   return (
     <div className='Home'>
         <div className='title'>Usuarios</div>
-        <Form />
-        <TableComponent data={usersList} />
-        {/* <button onClick={getAllUsers}>Listar Usuarios</button> */}
+        <UserForm />
+        <UsersTable data={usersList} />
     </div>
   )
 }
